refactor: update error and warning colors
Update the colors used for both warnings and errors, especially the
background and border colors to make it easier to read on both dark and
light themes.

- For the light theme both the yellow-600 and red-600 were used, as well
as the bg color for the light theme. - For the dark theme both the
yellow-400 and red-400 were used, as well as the bg color for the dark
theme