chore: update zed theme with flexoki mappings
This commit introduces a lot of changes to Flexoki's Zed theme in light
of comparing each of the theme classes in the theme to Flexoki's
mappings – https://stephango.com/flexoki#mappings - for both the light
and dark version.

I tried to follow the mappings as much as possible, while also seeking
inspiration in Zed's One theme, which is the default theme that ships
with Zed and is maintained by Zed's team. Some inspiration was also
taken from Flexoki's VSCode screenshots to try and keep it as close to
that port as possible, when in doubt.

It's worth noting that for both dark and light versions,
`terminal.ansi.dim_black` uses the corresponding `tx-2` color because
that is the foreground of the color shown on temrinal's
auto-completion, so we need to make sure it is not using the `tx`
color by default and that it correctly stands out as a suggestion.